Add explosion resistance
