+++ GTF_Xp/BepInExLoader.cs
@@ -26,7 +26,7 @@ namespace GTFuckingXP
         MODNAME = "GTFuckingXP",
         AUTHOR = "Endskill",
         GUID = AUTHOR + "." + MODNAME,
-        VERSION = "2.2.1";
+        VERSION = "2.3.0";
 
         public static bool RundownDevMode { get; private set; }
         public static ConfigEntry<bool> DebugMessages { get; private set; }
@@ -92,7 +92,7 @@ namespace GTFuckingXP
             //General xp mod usage
             Harmony.PatchAll(typeof(GS_InLevelPatches));
             Harmony.PatchAll(typeof(EnemyDamageLimbPatches));
-            Harmony.PatchAll(typeof(PlayerFireDamagePatches));
+            Harmony.PatchAll(typeof(PlayerDamagePatches));
             Harmony.PatchAll(typeof(SentryGunCheckPatches));
             Harmony.PatchAll(typeof(SentryGunFiringPatches));
             Harmony.PatchAll(typeof(GS_AfterLevelPatches));
